requirements.txt
joblib==1.1.1
numpy==1.23.5
Pillow==9.5.0
pandas==2.2.1
seaborn==0.13.1
matplotlib==3.4.1
streamlit==1.31.0
scikit-learn
xgboost